+++ examples/cmake/mpu6500/drdy_spi.cc
@@ -23,59 +23,35 @@
 * IN THE SOFTWARE.
 */
 
-#include "mpu9250.h"
+#include "mpu6500.h"
 
-/* Mpu9250 object, SPI bus, CS on pin 10 */
-bfs::Mpu9250 imu(&SPI, 10);
-/* Sample rate divider driven by IMU data ready interrupt at 1000 Hz */
-volatile int srd = 0;
-/* Arrays for copying data */
-int8_t fifo_len;
-float ax_mps2[bfs::Mpu9250::FIFO_MAX_SIZE()];
-float ay_mps2[bfs::Mpu9250::FIFO_MAX_SIZE()];
-float az_mps2[bfs::Mpu9250::FIFO_MAX_SIZE()];
-float gx_radps[bfs::Mpu9250::FIFO_MAX_SIZE()];
-float gy_radps[bfs::Mpu9250::FIFO_MAX_SIZE()];
-float gz_radps[bfs::Mpu9250::FIFO_MAX_SIZE()];
+/* Mpu6500 object, SPI bus, CS on pin 10 */
+bfs::Mpu6500 imu(&SPI, 10);
 
 /* Data acquisition ISR */
 void imu_isr() {
-  srd++;
-  /* Collect FIFO data at 50 Hz */
-  if (srd > 19) {
-    srd = 0;
-    /* Read the FIFO buffer */
-    fifo_len = imu.ReadFifo();
-    imu.fifo_accel_x_mps2(ax_mps2, bfs::Mpu9250::FIFO_MAX_SIZE());
-    imu.fifo_accel_y_mps2(ay_mps2, bfs::Mpu9250::FIFO_MAX_SIZE());
-    imu.fifo_accel_z_mps2(az_mps2, bfs::Mpu9250::FIFO_MAX_SIZE());
-    imu.fifo_gyro_x_radps(gx_radps, bfs::Mpu9250::FIFO_MAX_SIZE());
-    imu.fifo_gyro_y_radps(gy_radps, bfs::Mpu9250::FIFO_MAX_SIZE());
-    imu.fifo_gyro_z_radps(gz_radps, bfs::Mpu9250::FIFO_MAX_SIZE());
-    /*
-    * Print the FIFO length (should be 20) and the last FIFO sample for all
-    * channels as an example. This FIFO buffer could be used to compute delta
-    * theta and delta velocity using 1000 Hz data, while having the
-    * microcontroller sample the IMU at a much lower rate.
-    */
-    Serial.print(fifo_len);
+  /* Check if data read */
+  if (imu.Read()) {
+    Serial.print(imu.new_imu_data());
     Serial.print("\t");
-    Serial.print(ax_mps2[fifo_len - 1]);
+    Serial.print(imu.accel_x_mps2());
     Serial.print("\t");
-    Serial.print(ay_mps2[fifo_len - 1]);
+    Serial.print(imu.accel_y_mps2());
     Serial.print("\t");
-    Serial.print(az_mps2[fifo_len - 1]);
+    Serial.print(imu.accel_z_mps2());
     Serial.print("\t");
-    Serial.print(gx_radps[fifo_len - 1]);
+    Serial.print(imu.gyro_x_radps());
     Serial.print("\t");
-    Serial.print(gy_radps[fifo_len - 1]);
+    Serial.print(imu.gyro_y_radps());
     Serial.print("\t");
-    Serial.print(gz_radps[fifo_len - 1]);
+    Serial.print(imu.gyro_z_radps());
+    Serial.print("\t");
+    Serial.print(imu.die_temp_c());
     Serial.print("\n");
   }
 }
 
-void setup() {
+int main() {
   /* Serial to display data */
   Serial.begin(115200);
   while(!Serial) {}
@@ -86,18 +62,17 @@ void setup() {
     Serial.println("Error initializing communication with IMU");
     while(1) {}
   }
+  /* Set the sample rate divider */
+  if (!imu.ConfigSrd(19)) {
+    Serial.println("Error configured SRD");
+    while(1) {}
+  }
   /* Enabled data ready interrupt */
   if (!imu.EnableDrdyInt()) {
     Serial.println("Error enabling data ready interrupt");
     while(1) {}
   }
-  /* Enable FIFO */
-  if (!imu.EnableFifo()) {
-    Serial.println("Error enabling FIFO");
-    while(1) {}
-  }
   /* Attach data ready interrupt to pin 9 */
   attachInterrupt(9, imu_isr, RISING);
+  while(1) {}
 }
-
-void loop() {}