+++ src/mpu6500.h
@@ -23,8 +23,8 @@
 * IN THE SOFTWARE.
 */
 
-#ifndef INVENSENSE_IMU_SRC_MPU9250_H_  // NOLINT
-#define INVENSENSE_IMU_SRC_MPU9250_H_
+#ifndef INVENSENSE_IMU_SRC_MPU6500_H_  // NOLINT
+#define INVENSENSE_IMU_SRC_MPU6500_H_
 
 #if defined(ARDUINO)
 #include <Arduino.h>
@@ -39,7 +39,7 @@
 
 namespace bfs {
 
-class Mpu9250 {
+class Mpu6500 {
  public:
   /* Sensor and filter settings */
   enum I2cAddr : uint8_t {
@@ -80,10 +80,10 @@ class Mpu9250 {
     WOM_RATE_250HZ = 0x0A,
     WOM_RATE_500HZ = 0x0B
   };
-  Mpu9250() {}
-  Mpu9250(TwoWire *i2c, const I2cAddr addr) :
+  Mpu6500() {}
+  Mpu6500(TwoWire *i2c, const I2cAddr addr) :
           imu_(i2c, static_cast<uint8_t>(addr)) {}
-  Mpu9250(SPIClass *spi, const uint8_t cs) :
+  Mpu6500(SPIClass *spi, const uint8_t cs) :
           imu_(spi, cs) {}
   void Config(TwoWire *i2c, const I2cAddr addr);
   void Config(SPIClass *spi, const uint8_t cs);
@@ -98,8 +98,6 @@ class Mpu9250 {
   inline uint8_t srd() const {return srd_;}
   bool ConfigDlpfBandwidth(const DlpfBandwidth dlpf);
   inline DlpfBandwidth dlpf_bandwidth() const {return dlpf_bandwidth_;}
-  bool EnableWom(int16_t threshold_mg, const WomRate wom_rate);
-  void Reset();
   bool Read();
   inline bool new_imu_data() const {return new_imu_data_;}
   inline float accel_x_mps2() const {return accel_[0];}
@@ -108,17 +106,13 @@ class Mpu9250 {
   inline float gyro_x_radps() const {return gyro_[0];}
   inline float gyro_y_radps() const {return gyro_[1];}
   inline float gyro_z_radps() const {return gyro_[2];}
-  inline bool new_mag_data() const {return new_mag_data_;}
-  inline float mag_x_ut() const {return mag_[0];}
-  inline float mag_y_ut() const {return mag_[1];}
-  inline float mag_z_ut() const {return mag_[2];}
   inline float die_temp_c() const {return temp_;}
 
  private:
   InvensenseImu imu_;
   int32_t spi_clock_;
   /*
-  * MPU-9250 supports an SPI clock of 1 MHz for config and 20 MHz for reading
+  * MPU-6500 supports an SPI clock of 1 MHz for config and 20 MHz for reading
   * data; however, in testing we found that 20 MHz was sometimes too fast and
   * scaled this down to 15 MHz, which consistently worked well.
   */
@@ -132,26 +126,19 @@ class Mpu9250 {
   float gyro_scale_, requested_gyro_scale_;
   uint8_t srd_;
   static constexpr float TEMP_SCALE_ = 333.87f;
-  uint8_t asa_buff_[3];
-  float mag_scale_[3];
   uint8_t who_am_i_;
-  static constexpr uint8_t WHOAMI_MPU9250_ = 0x71;
-  static constexpr uint8_t WHOAMI_MPU9255_ = 0x73;
-  static constexpr uint8_t WHOAMI_AK8963_ = 0x48;
+  static constexpr uint8_t WHOAMI_MPU6500_ = 0x70;
   /* Data */
   static constexpr float G_MPS2_ = 9.80665f;
   static constexpr float DEG2RAD_ = 3.14159265358979323846264338327950288f /
                                     180.0f;
-  bool new_imu_data_, new_mag_data_;
-  bool mag_sensor_overflow_;
-  uint8_t mag_data_[8];
-  uint8_t data_buf_[23];
-  int16_t accel_cnts_[3], gyro_cnts_[3], temp_cnts_, mag_cnts_[3];
-  float accel_[3], gyro_[3], mag_[3];
+  bool new_imu_data_;
+  uint8_t data_buf_[15];
+  int16_t accel_cnts_[3], gyro_cnts_[3], temp_cnts_;
+  float accel_[3], gyro_[3];
   float temp_;
   /* Registers */
   static constexpr uint8_t PWR_MGMNT_1_ = 0x6B;
-  static constexpr uint8_t H_RESET_ = 0x80;
   static constexpr uint8_t CLKSEL_PLL_ = 0x01;
   static constexpr uint8_t WHOAMI_ = 0x75;
   static constexpr uint8_t ACCEL_CONFIG_ = 0x1C;
@@ -166,51 +153,12 @@ class Mpu9250 {
   static constexpr uint8_t INT_RAW_RDY_EN_ = 0x01;
   static constexpr uint8_t INT_STATUS_ = 0x3A;
   static constexpr uint8_t RAW_DATA_RDY_INT_ = 0x01;
-  static constexpr uint8_t USER_CTRL_ = 0x6A;
-  static constexpr uint8_t I2C_MST_EN_ = 0x20;
-  static constexpr uint8_t I2C_MST_CLK_ = 0x0D;
-  static constexpr uint8_t I2C_MST_CTRL_ = 0x24;
-  static constexpr uint8_t I2C_SLV0_ADDR_ = 0x25;
-  static constexpr uint8_t I2C_SLV0_REG_ = 0x26;
-  static constexpr uint8_t I2C_SLV0_CTRL_ = 0x27;
-  static constexpr uint8_t I2C_SLV0_DO_ = 0x63;
-  static constexpr uint8_t I2C_READ_FLAG_ = 0x80;
-  static constexpr uint8_t I2C_SLV0_EN_ = 0x80;
-  static constexpr uint8_t EXT_SENS_DATA_00_ = 0x49;
-  /* Needed for WOM */
-  static constexpr uint8_t INT_WOM_EN_ = 0x40;
-  static constexpr uint8_t PWR_MGMNT_2_ = 0x6C;
-  static constexpr uint8_t DISABLE_GYRO_ = 0x07;
-  static constexpr uint8_t MOT_DETECT_CTRL_ = 0x69;
-  static constexpr uint8_t ACCEL_INTEL_EN_ = 0x80;
-  static constexpr uint8_t ACCEL_INTEL_MODE_ = 0x40;
-  static constexpr uint8_t LP_ACCEL_ODR_ = 0x1E;
-  static constexpr uint8_t WOM_THR_ = 0x1F;
-  static constexpr uint8_t PWR_CYCLE_WOM_ = 0x20;
-  /* AK8963 registers */
-  static constexpr uint8_t AK8963_I2C_ADDR_ = 0x0C;
-  static constexpr uint8_t AK8963_ST1_ = 0x02;
-  static constexpr uint8_t AK8963_DATA_RDY_INT_ = 0x01;
-  static constexpr uint8_t AK8963_HXL_ = 0x03;
-  static constexpr uint8_t AK8963_CNTL1_ = 0x0A;
-  static constexpr uint8_t AK8963_PWR_DOWN_ = 0x00;
-  static constexpr uint8_t AK8963_CNT_MEAS1_ = 0x12;
-  static constexpr uint8_t AK8963_CNT_MEAS2_ = 0x16;
-  static constexpr uint8_t AK8963_FUSE_ROM_ = 0x0F;
-  static constexpr uint8_t AK8963_CNTL2_ = 0x0B;
-  static constexpr uint8_t AK8963_RESET_ = 0x01;
-  static constexpr uint8_t AK8963_ASA_ = 0x10;
-  static constexpr uint8_t AK8963_WHOAMI_ = 0x00;
-  static constexpr uint8_t AK8963_HOFL_ = 0x08;
   /* Utility functions */
   bool WriteRegister(const uint8_t reg, const uint8_t data);
   bool ReadRegisters(const uint8_t reg, const uint8_t count,
                      uint8_t * const data);
-  bool WriteAk8963Register(const uint8_t reg, const uint8_t data);
-  bool ReadAk8963Registers(const uint8_t reg, const uint8_t count,
-                           uint8_t * const data);
 };
 
 }  // namespace bfs
 
-#endif  // INVENSENSE_IMU_SRC_MPU9250_H_ NOLINT
+#endif  // INVENSENSE_IMU_SRC_MPU6500_H_ NOLINT