+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v6.0.0
+- Including support for the MPU9250 adn MPU6500
+- Removed Eigen and Units dependencies, so it should work with any Arduino
+- Removed FIFO; there was some bug with the FIFO and no time currently to solve
+
 ## v5.6.0
 - Instead of inputting an I2C address, use an enum of the two potential addresses